Separate matchmaker processing functions
